Added function of title and tried to modify the fly
Technically I tried to add wings to the fly but got overwhelmed, so I proceed to explore and try adding a title screen.
Good news: I succeed creating the function for the title + I used let and if
Bad news: It's not showing up on the simulation.
I'll continue trying!
(Also add some stuff in the ideas )